+++ cli/build.rs
@@ -362,8 +362,7 @@ fn create_cli_snapshot(snapshot_path: PathBuf) {
     deno_io::deno_io::init_ops(Default::default()),
     deno_fs::deno_fs::init_ops::<PermissionsContainer>(false),
     deno_flash::deno_flash::init_ops::<PermissionsContainer>(false), // No --unstable
-    deno_node::deno_node_loading::init_ops::<PermissionsContainer>(None), // No --unstable.
-    deno_node::deno_node::init_ops(),
+    deno_node::deno_node::init_ops::<PermissionsContainer>(None),
     cli::init_ops_and_esm(), // NOTE: This needs to be init_ops_and_esm!
   ];
