+++ cli/tools/repl/session.rs
@@ -507,6 +507,7 @@ impl ReplSession {
         deno_node::initialize_runtime(
           &mut self.worker.js_runtime,
           self.proc_state.options.has_node_modules_dir(),
+          None,
         )?;
         self.has_initialized_node_runtime = true;
       }
